+++ src/index.d.ts
@@ -1,11 +1,22 @@
 /**
  * Used in Autocomplete and Select components
  */
-export interface OptionsType {
+export interface SelectOptionsType {
   label: string;
   value: string | number;
 }
 
+/**
+ * Used in RadioGroup component
+ */
+export interface RadioGroupOptionsType {
+  label: string;
+  value: string;
+}
+
 export * from './Autocomplete'
+export * from './DateTimePicker'
+export * from './RadioGroup'
 export * from './Select'
+export * from './Switch'
 export * from './TextField'