+++ src/index.js
@@ -1,4 +1,6 @@
-
 export * from './Autocomplete'
+export * from './DateTimePicker'
+export * from './RadioGroup'
 export * from './Select'
+export * from './Switch'
 export * from './TextField'